:recycle: rename to text_color
