+++ lib/initial_setting/initial_setting.dart
@@ -2,7 +2,7 @@ import 'package:Pilll/theme/color.dart';
 import 'package:Pilll/initial_setting/initial_setting_1.dart';
 import 'package:Pilll/model/pill_sheet_type.dart';
 import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_style.dart';
+import 'package:Pilll/theme/text_color.dart';
 import 'package:flutter/material.dart';
 
 typedef InitialSettingCallback = void Function(InitialSettingModel);
