+++ lib/initial_setting/initial_setting2.dart
@@ -3,7 +3,7 @@ import 'package:Pilll/initial_setting/initial_setting_3.dart';
 import 'package:Pilll/main/record/pill_sheet.dart';
 import 'package:Pilll/main/record/pill_sheet_model.dart';
 import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_style.dart';
+import 'package:Pilll/theme/text_color.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
