+++ lib/initial_setting/initial_setting_4.dart
@@ -1,7 +1,7 @@
 import 'package:Pilll/theme/color.dart';
 import 'package:Pilll/initial_setting/initial_setting.dart';
 import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_style.dart';
+import 'package:Pilll/theme/text_color.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
